# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SyRide API",
        "description": "SyRide Ride-Sharing Platform – Full API Reference",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/4th_year_project/public",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/profile/verify/passenger": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Submit passenger verification request",
                "operationId": "verifyPassenger",
                "responses": {
                    "200": {
                        "description": "Verification submitted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/verify/driver": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Submit driver verification request",
                "operationId": "verifyDriver",
                "responses": {
                    "200": {
                        "description": "Verification submitted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}